requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
annotated-types==0.6.0
Appium-Python-Client==3.1.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
future==0.18.3
h11==0.14.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
outcome==1.3.0.post0
packaging==23.2
pluggy==1.3.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
PySocks==1.7.1
pytest==7.4.3
python-dotenv==1.0.0
requests==2.31.0
selene==2.0.0rc4
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==2.0.7
webdriver-manager==3.8.6
wsproto==1.2.0
